Add day 4, update day 3 test to ignore parts, clean up day 2
Oof.  That's all I can say on this one.  Going into this one, I didn't
read the problem closely, so ended up going at it entirely wrong.
Luckily, wasn't much to salvage it, but the method is still a mess
overall.  Gets the right answer though and is doesn't take too long to
do it either, so that's good.  Can't say the same for Day03, hence I
added the ignores to that one.  Was adding ~3s to the tests.

Cleaned up Day02 a bit.  The method for scoring the strings was a bit
over complicated for what the question was asking.  No need for the
length check, as they are all the same length.  Also, the `map + sum`,
wasn't necessary, as a count would have worked just as well.