Add day 05
First look at this challenge, I thought the answer was going to be
rather straightforward.  In the end, was a bit more complex, leading
to yet another messy method.  Luckily, like the previous, both parts
could use the same code, so that cut down on the work needed.
Unfortunately, the solution for the second part was basically
brute-force, so it takes ~2min.  Obviously ignored now in the tests